membersihkan staging

+++ app/src/main/java/com/finpro/garudanih/view/detils/DetailPesawatActivity.kt
@@ -1,17 +1,15 @@
 package com.finpro.garudanih.view.detils
 
 import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import android.widget.Toast
-import androidx.activity.viewModels
+import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
 import com.finpro.garudanih.databinding.ActivityDetailPesawatBinding
 import com.finpro.garudanih.model.Ticket
+import com.finpro.garudanih.view.HomeBottomActivity
 import com.finpro.garudanih.view.pemesanan.PemesananActivity
 import com.finpro.garudanih.viewmodel.TiketViewModel
-import com.finpro.garudanih.viewmodel.UserViewModel
 import com.finpro.garudanih.wishlist.DataWishPesawatLoc
 import com.finpro.garudanih.wishlist.DatabaseWishPesawatLoc
 import com.finpro.garudanih.wishlist.WishPesawatDaoLoc
@@ -77,7 +75,7 @@ class DetailPesawatActivity : AppCompatActivity() {
         binding.wishlist.setOnClickListener{
             GlobalScope.async {
                 val d = intent.getSerializableExtra("detail") as Ticket
-                val idd = detail.id.toInt()
+                val idd = detail.id
                 val asal = detail.destinationCode
                 val tujuan = detail.departureCode
                 val jad = detail.takeOff
@@ -100,28 +98,5 @@ class DetailPesawatActivity : AppCompatActivity() {
             }
         }
     }
-
-//    fun getListPesawat(){
-//        val itemListPesawat = intent
-//
-//        val kota = itemListPesawat.getStringExtra("destinasi")
-//        val keberangkatan = itemListPesawat.getStringExtra("departure")
-//        val jadwal = itemListPesawat.getStringExtra("jadwal")
-//        val harga = itemListPesawat.getIntExtra("harga",0)
-//        val chair = itemListPesawat.getIntExtra("totalchair",0)
-//        val status = itemListPesawat.getStringExtra("class")
-////        val image = itemListPesawat.getIntExtra("image",0)
-//
-//        binding.txtInputAsal.text = keberangkatan
-//        binding.txtInputTujuan.text = kota
-//        binding.txtHargaDetail.text = "Harga Tiket \nRp"+harga.toString()
-//        binding.txtJadwal.text = "Jadwal : \n"+jadwal
-//        binding.txtChair.text =  "Available "+chair.toString()
-//        binding.txtClass.text = status+" Class"
-////        binding.ivKota.setImageResource(image)
-//
-//
-//
-//    }
 
 }
